add multi-class support

+++ src/io/DataPoint.h
@@ -139,7 +139,7 @@ namespace BOC {
 		void Normalize(){
 			if (this->sum_sq != 0){
 				size_t featNum = this->indexes.size();
-				for (int i = 0; i < featNum; ++i){
+				for (size_t i = 0; i < featNum; ++i){
 					this->features[i] /= this->sum_sq;
 				}
 			}
